Update utility_functions.R
Now interpolation really works.
It has to be tested when the retention time is not sampled at constat sampling frequency.

A side effect of interpoaltion is an increase on the size of data matrix. To be studied: force data matrix to be a matrix of integers as happens with the original matrix