Performance improvement
realize.GCIMSDataset passes all delayed_ops to each worker. This means that we want to keep delayed_ops as small as possible.

A delayed op is defined with several functions. A function is defined by (a) the body of the function [code], (b) the formals [arguments] and (c) the environment [context].

The environment limits what variables are visible for a function.

Compare these two implementations:

slow_impl <- function() {
    large_object <- runif(1E6)
    myfun_slow <- function(x) x
    BiocParallel::bplapply(
        1:10,
        myfun_slow
    )
}

myfun_fast <- function(x) x
fast_impl <- function() {
    large_object <- runif(1E6)
    BiocParallel::bplapply(
        1:10,
        myfun_fast
    )
}

BiocParallel must pass in both cases myfun to the workers, so they can evaluate it. myfun_slow has access to large_object through its enclosing environment, so BiocParallel may have to pass it to the worker, with the corresponding overhead.

myfun_fast does not have access to large_object so its environment is smaller to send.